R: rename runtimecontext to cfcontext
