+++ cloudflare/internal/cfruntimecontext/cfruntimecontext.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"errors"
 	"syscall/js"
-
-	"github.com/syumai/workers/internal/runtimecontext"
 )
 
 /**
@@ -49,7 +47,7 @@ var ErrValueNotFound = errors.New("execution context value for specified key not
 // GetRuntimeContextValue gets value for specified key from RuntimeContext.
 // - if the value is undefined, return error.
 func GetRuntimeContextValue(ctx context.Context, key string) (js.Value, error) {
-	runtimeCtxValue := runtimecontext.MustExtract(ctx)
+	runtimeCtxValue := cfcontext.MustExtract(ctx)
 	v := runtimeCtxValue.Get(key)
 	if v.IsUndefined() {
 		return js.Value{}, ErrValueNotFound
